Move forkId tests to forkId package
